pipeline/requirements.txt
requests>=2.28
icalendar>=5.0
python-dateutil>=2.8
